# Cache optimized regexp matchers